add revocation ids to the specification
we generate unique revocation ids for each block of the token, that also
depend on the previous block. That way, when holding a token, we also
have valid revocation ids for each of its parent tokens.

Since they are generated from hashes of the token's data, revocation can
be added to an existing system even after tokens were created