restrict the execution scope to the authority and current block (#99)
* add read, write and time to the default symbol table

remove authority an ambient from the symbol table, as they are not used
anymore

* restrict the execution scope to the authority and current block

previously when evaluating a block, we used facts from the authorizer
and all previous blocks. This changes the execution scope to use only
the authority block, authorizer and the current block, which is a much
safer default and aligns more with expectations.

Execution using all previous blocks will be accessible again in the
future through an option